Remove default server URL
The usage of https://tuleap.net is very circumstantial as it only makes
sense for the users of Tuleap.net playing CI jobs (i.e. mainly Tuleap core
devs). Virtually everyone will be forced to set the server's URL anyway.

From past experiences, it might be better to have an hard crash instead of
something that tries to query a real Tuleap instance. Debugging and understanding
of what the issue is gets easier. Also, from a privacy point of view it makes
sure users of the library does not leak the existence (and more, user agents are
verbose) of their Jenkins instance inadvertently.